Fix build errors on Clang
[CL 23020991 by charles lefebvre in ue5-main branch]

+++ Engine/Source/Runtime/AIModule/Classes/BehaviorTree/BlackboardData.h
@@ -86,7 +86,7 @@ public:
 
 	FORCEINLINE FBlackboard::FKey GetFirstKeyID() const { return FirstKeyID; }
 
-	bool IsValidKey(FBlackboard::FKey KeyID) const { return KeyID != FBlackboard::InvalidKey && KeyID < GetNumKeys(); }
+	bool IsValidKey(FBlackboard::FKey KeyID) const { return KeyID != FBlackboard::InvalidKey && (int32)KeyID < GetNumKeys(); }
 
 	/** @return key data */
 	const FBlackboardEntry* GetKey(FBlackboard::FKey KeyID) const;
